data prep and exp separation, data saved into cvs
